+++ examples/as_f64.rs
@@ -18,14 +18,37 @@ type Result<T> = std::result::Result<T, Box<dyn std::error::Error>>;
 fn main() -> Result<()> {
     sensible_env_logger::init!();
 
+    trace!("With Empty String:");
     let data = r#"
     {
-        "timestamp": 123456789076543210
+        "timestamp": ""
     }"#;
 
     let m: Msg = serde_json::from_str(data).unwrap();
+    assert_eq!(m.timestamp, 0.0);
+    trace!("  {m:?}");
 
-    trace!("{m:?}");
+    trace!("With Null:  ");
+    let data = r#"
+    {
+        "timestamp": null
+    }"#;
+
+    let m: Msg = serde_json::from_str(data).unwrap();
+    assert_eq!(m.timestamp, 0.0);
+    trace!("  {m:?}");
+
+    trace!("With F64:");
+
+    let data = r#"
+    {
+        "timestamp": 123456789.076543210
+    }"#;
+
+    let m: Msg = serde_json::from_str(data).unwrap();
+    trace!("  {m:?}");
+
+    trace!("With String:");
 
     let data = r#"
     {
@@ -33,8 +56,19 @@ fn main() -> Result<()> {
     }"#;
 
     let m: Msg = serde_json::from_str(data).unwrap();
+    trace!("  {m:?}");
 
-    trace!("{m:?}");
+    trace!("With U64:");
+
+    let data = r#"
+    {
+        "timestamp": 123456789076543210
+    }"#;
+
+    let m: Msg = serde_json::from_str(data).unwrap();
+    trace!("  {m:?}");
+
+    trace!("With I64:");
 
     let data = r#"
     {
@@ -42,8 +76,7 @@ fn main() -> Result<()> {
     }"#;
 
     let m: Msg = serde_json::from_str(data).unwrap();
-
-    trace!("{m:?}");
+    trace!("  {m:?}");
 
     Ok(())
 }