+++ examples/as_u64.rs
@@ -18,14 +18,48 @@ type Result<T> = std::result::Result<T, Box<dyn std::error::Error>>;
 fn main() -> Result<()> {
     sensible_env_logger::init!();
 
+    trace!("With Empty String:");
+    let data = r#"
+    {
+        "timestamp": ""
+    }"#;
+
+    let m: Msg = serde_json::from_str(data).unwrap();
+    assert_eq!(m.timestamp, 0);
+    trace!("  {m:?}");
+
+    trace!("With Null:  ");
+    let data = r#"
+    {
+        "timestamp": null
+    }"#;
+
+    let m: Msg = serde_json::from_str(data).unwrap();
+    assert_eq!(m.timestamp, 0);
+    trace!("  {m:?}");
+
+    trace!("With U64:  ");
+
     let data = r#"
     {
         "timestamp": 123456789076543210
     }"#;
 
     let m: Msg = serde_json::from_str(data).unwrap();
+    trace!("  {m:?}");
 
-    trace!("{m:?}");
+    trace!("With F64:");
+
+    let data = r#"
+    {
+        "timestamp": 0.5
+    }"#;
+
+    let m: Msg = serde_json::from_str(data).unwrap();
+    trace!("  {m:?}");
+    assert_eq!(m.timestamp, 1);
+
+    trace!("With String:  ");
 
     let data = r#"
     {
@@ -33,8 +67,9 @@ fn main() -> Result<()> {
     }"#;
 
     let m: Msg = serde_json::from_str(data).unwrap();
+    trace!("  {m:?}");
 
-    trace!("{m:?}");
+    trace!("With I64:  ");
 
     let data = r#"
     {
@@ -42,9 +77,9 @@ fn main() -> Result<()> {
     }"#;
 
     if let Err(e) = serde_json::from_str::<Msg>(data) {
-        trace!("Expected error: {}", e);
+        trace!("  Expected error: {}", e);
     } else {
-        error!("ERROR! An overflow error should have occurred.");
+        error!("  ERROR! An overflow error should have occurred.");
         assert_eq!(0, 1, "failure!");
     };